requirements.txt
aniso8601==7.0.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
promise==2.2.1
pytz==2019.2
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.4
